requirements.txt
streamlit
camelot-py[cv]
sentence-transformers
